requirements.txt
boto3==1.35.11
botocore==1.35.11
click==8.1.7
colorama==0.4.6
cramjam==2.8.3
fastparquet==2024.5.0
fsspec==2024.6.1
jmespath==1.0.1
joblib==1.4.2
lightgbm==4.5.0
nltk==3.9.1
numpy==2.1.1
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.2
